Fixed an issue in which full paths wouldn't work properly
On Content.cs (I'm guessing by mistake) there was an addition of 2 strings that form a path. However, when the path provided is a full path then it doesn't make sens cause you end up with a full path followed by another full path. Using Path.Combine the standard library this case is taken care of.
```c#
rootPath + item.Path -> Path.Combine(rootPath, item.Path)
```
As for the other change in NPLConfigReader.cs, if the path provided is a full path instead of a relative path, then we don't really want to remove anything as it is already a correct path. So If the path is rooted just use the path as is:
```c#
string name = file;
if (!Path.IsPathRooted(file))
     name = file.Remove(0, rootDir.Length).Replace('\\', '/');
var newItem = new Item(name);
```